Readme.md
How to Use ?
Enter a Number for a specific output:

(1) : To show all available tools and type the number of a tool which you want to install.
(2) : To show tools category.
(3) : If you want to update Tool-X.
(4) : If you want to know About Us.
(x) : To exit the tool.

Type Tool-X or toolx from anywhere in your terminal to open Tool-X.

How to remove ?
Open Tool-X and type rm -T or uninstall tool-x to uninstall Tool-X.

Warning

We are not responsible for any misuse or damage caused by this program. use this tool at your own risk!